Move `ByteArrays` to utils
